Added getOrderInfo function and updated submitOrder in orderControllers.js
I introduced a new function called `getOrderInfo` to handle GET requests for fetching user orders. This function decrypts the card information before sending it to the front-end, using the `decrypt` function that's also used elsewhere in the application for cardInfo.

I also made changes to the `submitOrder` function. Now, it takes an estimated delivery date as an input and includes it in both the database entry for the new order and the order confirmation email.