Integrated UserContext into OrderSummary Component
In this update, I imported UserContext into the OrderSummary component. The key change is the use of the setOrderInfo function from UserContext. This allows the app to instantly update the order information in the context as soon as a new order is placed. This way, the user sees their most recent orders without needing to refresh the page or make extra API calls.