Modified the createCardInfo function in the billingInfoControllers.js
In the old version of createCardInfo, the function encrypted the card data and stored it in the database, then sent this encrypted data back in the response. This was not useful for front-end operations that required the card's original data.

In the new version, I made two card objects: encryptedCard for database storage and unencryptedCard for the front-end. The unencryptedCard includes all the original card details and also the new card's ID from the database. This unencrypted card object is what gets sent back in the response. This change allows the front-end to receive the necessary data in its original, unencrypted form.

This fixed a bug that I've been experiencing for a long time now. Before this change, a user creates a new card and that new card gets added to the context but it was missing it's id so it wouldn't function until the user refreshed the page. At that point the full details of the new card were retrieved and the user could use the new card. This change fixed that issue and no refresh is required!