Added Partial Guest Checkout Support and GuestModal Component
I made several changes to my React e-commerce application to better handle scenarios where a user who is not logged in (a guest) navigates to the checkout page. Although this feature is not yet fully functional, progress has been made.

Initially, I incorporated checks within the CheckoutDetails component and introduced new state variables tailored for guest checkout. I also developed a new component, along with a back-end controller function and routes, to facilitate the creation of a guest user in the database.

After some deliberation, I opted to temporarily suspend this development. In its place, I implemented a GuestModal component that activates when a guest tries to access the checkout page. This modal advises them to either create an account or log in to utilize the checkout feature fully.

Given the extended development timeline of this application, my current priority has shifted towards deploying the website. I intend to revisit and finalize this guest checkout feature at a later date.