Implemented Enhanced Rate Limiting for Contact Us Route
As part of the final preparations for deploying my React e-commerce application, I introduced a new, more stringent rate limiter using the express-rate-limit library. This enhanced rate limiter, named strictLimit, is currently applied exclusively to the Contact Us route.

The strictLimit configuration allows only 2 requests per IP within a 15-minute window. If this limit is exceeded, the server responds with a 429 status code and a message indicating that too many requests have been made. This is in contrast to the previous rate limiter, limit, which allowed up to 5 requests per IP in a 5-minute window.

The implementation of this stricter rate limiter serves a dual purpose. First, it helps to minimize unnecessary server load, which is crucial as the application is nearing its deployment phase. Second, it mitigates the risk of misuse, particularly for routes that trigger email sending via AWS SES—a service for which excessive usage could result in financial costs.

As for the SignUpModal component I added a new input validation for the email field that ensures the email is in the correct format.